added gRPC client libraries

+++ src/Services/Web/Business/Entities/Order.cs
@@ -5,7 +5,7 @@ namespace PixieFit.Web.Business.Entities;
 public class Order : BaseEntity
 {
     public Guid UserId { get; set; }
-    public long TotalAmount { get; set; }
+    public decimal TotalAmount { get; set; }
     public PaymentMethod PaymentMethod { get; set; }
     public PaymentStatus PaymentStatus { get; set; }
     public Guid CreditPackageId { get; set; }
